requirements.txt
Django==4.2.4
psycopg2-binary==2.9.1
djangorestframework==3.14.0
gunicorn==19.9.0
drf-yasg==1.21.7
enum34==1.1.10
django-enumchoicefield==3.0.1

mailjet-rest==1.3.4